examples/5_streamlit-chatbot/pyproject.toml
[project]
name = "stremlit-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.66.3",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "streamlit>=1.43.2",
]
